Use '/usr/local/bin/perl' instead of '/usr/bin/env perl'
Everywhere; all shebangs and examples.

This was chosen over using the C<#!/usr/bin/env perl> trick
because if you're in an environment where you need a local lib installation
you probably don't have the choice of installing new perls.
If there are in fact multiple perls available to you at installation time
it seems like more of a feature to link to a particular perl
than to use the system default.

It probably isn't a good idea
to use the same local lib for multiple perls, anyway.